refactor yaml alias scanner
